some functions moved from selection_functions to item_getters
