+++ functions/secondary/item_functions.py
@@ -7,12 +7,14 @@ try:  # Assume we're a submodule in a package.
     from base.classes.auto import AUTO, Auto
     from base.functions.arguments import update
     from content.items.item_type import ItemType
+    from content.items.item_getters import get_composite_key
     from content.selection import selection_functions as sf
     from functions.primary import items as it
 except ImportError:  # Apparently no higher-level package has been imported, fall back to a local import.
     from ...base.classes.auto import AUTO, Auto
     from ...base.functions.arguments import update
     from ...content.items.item_type import ItemType
+    from ...content.items.item_getters import get_composite_key
     from ...content.selection import selection_functions as sf
     from ..primary import items as it
 
@@ -29,7 +31,7 @@ def composite_key(*functions) -> Callable:
     key_functions = update(functions)
 
     def func(item) -> tuple:
-        return sf.get_composite_key(item=item, keys_descriptions=key_functions)
+        return get_composite_key(item=item, keys_descriptions=key_functions)
     return func
